remove garbage collection of mh_q and mh objects
there's a good reason for that. When an mh_q object is passed to an mh constructor,
the mh constructor takes ownership of the mh_q object, and so must be the only responsible
for mh_q destruction. In the same way, when an mh object is passed to the amh constructor,
the new amh objects takes ownership of the mh object. I dunno better way of handling that other than
requiring explicit free of mh_q and mh objects by the scheme users